requirements.txt
pandas
requests
beautifulsoup4
python-dotenv
openai
airflow